refactor: rename gas-snapshots to function-gas-snapshots
