+++ v-next/hardhat/src/internal/builtin-plugins/gas-analytics/function-gas-snapshots.ts
